Nighty build 100
More work done in MainBody. Num <= 2.01 case is almost complete.
Methods such as getups and reshapeDmatrix to streamline code. Made
various changes to many files to pass up data structures from deep
within nested constructors. Some redundant casting removed from Locate1
and LocateC1.

+++ src/Locate1.java
@@ -31,7 +31,7 @@ public class Locate1 {
 		i = new double[][]{{1,0},{0,0}};
 		j = new double[][]{{0,1},{0,0}};
 		k = new double[][]{{-1,1},{0,0}};
-		d = new double[][]{{(double) (q.get(0, 0)-p.get(0, 0)),(double) (q.get(0, 1)-p.get(0, 1))},{0,0}};
+		d = new double[][]{{(q.get(0, 0)-p.get(0, 0)),(q.get(0, 1)-p.get(0, 1))},{0,0}};
 		System.out.println(d[0][0]+" " +d[0][1]);
 		i1 = new Matrix(i);
 		j1 = new Matrix(j);
@@ -102,7 +102,7 @@ public class Locate1 {
 			Matrix c = p;
 			p=q;
 			q=c;
-			d = new double[][]{{(double) (q.get(0, 0)-p.get(0, 0)),(double) (q.get(0, 1)-p.get(0, 1))},{0,0}};
+			d = new double[][]{{(q.get(0, 0)-p.get(0, 0)),(q.get(0, 1)-p.get(0, 1))},{0,0}};
 			System.out.println(d[0][0]+" " +d[0][1]);
 			case1(p,q,H);
 			case2(p,q,H);
